client: Render directly from game state
The game now renders everything directly from the data structures
defined in the state module.  The target is to implement the whole game
logic in there and let it run on the server and the clients
simultaneously.  This enables us to interpolate between updates of the
game state from the server.